# specs/tsp-output/@typespec/openapi3/generated/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: Comprehensive API for managing beauty salon operations including salons, staff, services, customers, reservations, bookings, and reviews.
  version: 0.0.0
...
paths:
...
  /reviews:
    get:
      operationId: ReviewOperations_list
      summary: List reviews
      parameters:
        - name: salonId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.SalonId'
          explode: false
        - name: customerId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.CustomerId'
          explode: false
        - name: staffId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Models.StaffId'
          explode: false
        - name: minRating
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: isVerified
          in: query
          required: false
          schema:
            type: boolean
          explode: false
        - $ref: '#/components/parameters/Models.PaginationParams.limit'
        - $ref: '#/components/parameters/Models.PaginationParams.offset'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total
                  - limit
                  - offset
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Models.Review'
                  total:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  offset:
                    type: integer
                    format: int32
      tags:
        - Reviews
    post:
      operationId: ReviewOperations_create
      summary: Create review
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          headers:
            location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Review'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.CreateReviewRequest'
  /reviews/{id}:
    get:
      operationId: ReviewOperations_get
      summary: Get review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReviewId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Review'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reviews
    patch:
      operationId: ReviewOperations_update
      summary: Update review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReviewId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Review'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.UpdateReviewRequest'
    delete:
      operationId: ReviewOperations_delete
      summary: Delete review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReviewId'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reviews
  /reviews/{id}/helpful:
    post:
      operationId: ReviewOperations_markHelpful
      summary: Mark review as helpful
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.ReviewId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  helpfulCount:
                    type: integer
                    format: int32
                required:
                  - helpfulCount
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Reviews
...
  /salons/{salonId}/reviews:
    get:
      operationId: SalonReviewOperations_getSalonReviews
      summary: Get salon reviews
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.SalonId'
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum:
              - recent
              - rating
              - helpful
          explode: false
        - $ref: '#/components/parameters/Models.PaginationParams.limit'
        - $ref: '#/components/parameters/Models.PaginationParams.offset'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total
                  - limit
                  - offset
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Models.Review'
                  total:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  offset:
                    type: integer
                    format: int32
      tags:
        - Reviews
  /salons/{salonId}/reviews/summary:
    get:
      operationId: SalonReviewOperations_getSummary
      summary: Get salon reviews summary
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.SalonId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.ReviewSummary'
      tags:
        - Reviews
...
  /staff/{staffId}/reviews:
    get:
      operationId: StaffReviewOperations_getStaffReviews
      summary: Get staff reviews
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.StaffId'
        - $ref: '#/components/parameters/Models.PaginationParams.limit'
        - $ref: '#/components/parameters/Models.PaginationParams.offset'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total
                  - limit
                  - offset
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Models.Review'
                  total:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  offset:
                    type: integer
                    format: int32
      tags:
        - Reviews
...
components:
  parameters:
    Models.PaginationParams.limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 20
      explode: false
    Models.PaginationParams.offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 0
      explode: false
  schemas:
...
    Models.CreateReviewRequest:
      type: object
      required:
        - salonId
        - customerId
        - reservationId
        - rating
      properties:
        salonId:
          $ref: '#/components/schemas/Models.SalonId'
        customerId:
          $ref: '#/components/schemas/Models.CustomerId'
        reservationId:
          $ref: '#/components/schemas/Models.ReservationId'
        staffId:
          $ref: '#/components/schemas/Models.StaffId'
        rating:
          type: integer
          format: int32
        comment:
          type: string
        serviceRating:
          type: integer
          format: int32
        staffRating:
          type: integer
          format: int32
        atmosphereRating:
          type: integer
          format: int32
        images:
          type: array
          items:
            type: string
...
    Models.CustomerId:
      type: string
...
    Models.Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
    Models.ReservationId:
      type: string
...
    Models.Review:
      type: object
      required:
        - id
        - salonId
        - customerId
        - reservationId
        - rating
        - isVerified
        - helpfulCount
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Models.ReviewId'
        salonId:
          $ref: '#/components/schemas/Models.SalonId'
        customerId:
          $ref: '#/components/schemas/Models.CustomerId'
        reservationId:
          $ref: '#/components/schemas/Models.ReservationId'
        staffId:
          $ref: '#/components/schemas/Models.StaffId'
        rating:
          type: integer
          format: int32
        comment:
          type: string
        serviceRating:
          type: integer
          format: int32
        staffRating:
          type: integer
          format: int32
        atmosphereRating:
          type: integer
          format: int32
        images:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        helpfulCount:
          type: integer
          format: int32
          default: 0
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    Models.ReviewId:
      type: string
    Models.ReviewSummary:
      type: object
      required:
        - averageRating
        - totalReviews
        - ratingDistribution
      properties:
        averageRating:
          type: number
          format: float
        totalReviews:
          type: integer
          format: int32
        ratingDistribution:
          type: object
          additionalProperties:
            type: integer
            format: int32
...
    Models.SalonId:
      type: string
...
    Models.StaffId:
      type: string
...
    Models.UpdateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          format: int32
        comment:
          type: string
        serviceRating:
          type: integer
          format: int32
        staffRating:
          type: integer
          format: int32
        atmosphereRating:
          type: integer
          format: int32
        images:
          type: array
          items:
            type: string
      description: Review update request with optional fields for partial updates
...
servers:
  - url: https://api.beauty-salon.com
    description: Beauty Salon API
    variables: {}